Install the Flask-api and initialize it
